requirements-init.txt
git+https://github.com/daemon/argconf
easydict==1.9
torchtext==0.3.1
tqdm==4.35.0
pytorch_pretrained_bert
nltk==3.4.5
allennlp==0.9.0
